[FSDP()][7/N] Refactor most of ctor
[ghstack-poisoned]
