+++ torch/distributed/fsdp/_utils.py
@@ -10,8 +10,6 @@ from torch.nn.parallel.scatter_gather import (  # type: ignore[attr-defined]
 )
 from torch.nn.utils.rnn import PackedSequence
 
-FSDP_FLATTENED = "_fsdp_flattened"
-
 
 def _contains_batchnorm(module):
     return any(isinstance(mod, _BatchNorm) for mod in module.modules())
@@ -61,33 +59,6 @@ def _apply_to_tensors(
     return apply(container)
 
 
-def _apply_to_modules(
-    root_module: torch.nn.Module,
-    module_fn: Callable,
-    return_fn: Callable,
-    *args,
-    **kwargs,
-):
-    """
-    Performs a pre-order traversal of the modules in the hierarchy rooted at
-    ``root_module``, applying ``module_fn`` at each module and finally
-    returning a value using ``return_fn``. The traversal constructs the full
-    module prefix name (e.g. "module.submodule." just like in model state dict)
-    and makes that available to ``module_fn``.
-    """
-
-    def f(module: torch.nn.Module, prefix: str, *args, **kwargs):
-        # Call the module function before recursing over children (pre-order)
-        module_fn(module, prefix, *args, **kwargs)
-        for submodule_name, submodule in module.named_children():
-            if submodule is not None:
-                new_prefix = prefix + submodule_name + "."
-                f(submodule, new_prefix, *args, **kwargs)
-
-    f(root_module, "", *args, **kwargs)
-    return return_fn(*args, **kwargs)
-
-
 @torch.no_grad()
 def _alloc_storage(tensor: torch.Tensor, size: torch.Size) -> bool:
     """
@@ -128,19 +99,6 @@ def _free_storage(tensor: torch.Tensor) -> bool:
         )
         tensor.storage().resize_(0)
     return not already_freed
-
-
-def _set_fsdp_flattened(tensor: torch.Tensor) -> None:
-    """
-    Sets an attribute on ``tensor`` to mark it as flattened by FSDP. This is to
-    avoid re-flattening it during nested construction.
-    """
-    setattr(tensor, FSDP_FLATTENED, True)
-
-
-def _is_fsdp_flattened(tensor: torch.Tensor) -> bool:
-    """Returns if ``tensor`` has been marked as flattened by FSDP."""
-    return getattr(tensor, FSDP_FLATTENED, False)
 
 
 def _same_storage(x: torch.Tensor, y: torch.Tensor) -> bool:
